requirements.txt
pandas
azure-identity
azure-ai-projects
python-dotenv